+++ components/finalized/Finalized.js
@@ -6,32 +6,28 @@ import {
     TouchableOpacity,
     Modal,
     ActivityIndicator,
+    ScrollView,
+    SafeAreaView,
 } from "react-native";
 import { connect } from "react-redux";
 import * as actions from "../../actions";
 import Schedule from "./Schedule";
 import Map from "./Map";
-import TransitRoutes from "./TransitRoutes";
-import {
-    data_timeline,
-    genreEventObjectArray,
-} from "../../reusable-functions/data_timeline";
 import { MaterialCommunityIcons } from "@expo/vector-icons";
-import TransitRoute from "./TransitRoutes";
+import { YellowBox } from "react-native";
 
 const Finalized = (props) => {
+    YellowBox.ignoreWarnings(["VirtualizedLists should never be nested"]);
     const [isLoading, setIsLoading] = React.useState(true);
     const [visible, setVisible] = React.useState(false);
     const [coord, setCoord] = React.useState([]);
     const [data, setData] = React.useState([]);
-    const [routes, setRoutes] = React.useState([]);
 
     //const route = props.route.params.route;
     const accessRights = props.route.params.access;
     const weather = props.route.params.weather;
     const userGenres = props.route.params.userGenres;
     const allData = props.route.params.data;
-    //console.log(props.route.params.routeGuide);
 
     const onClose = () => {
         setVisible(false);
@@ -39,15 +35,6 @@ const Finalized = (props) => {
 
     const mapUpdate = (coord) => {
         setCoord(coord);
-        // setRoutes(routes)
-    };
-
-    const routeUpdate = (selected, unsatisfied) => {
-        let temp = routes;
-        const result = temp.map((item) => {
-            return item == unsatisfied.location ? selected.location : item;
-        });
-        setRoutes(result);
     };
 
     const weatherIcon = (weather) => {
@@ -79,7 +66,6 @@ const Finalized = (props) => {
         const passed = props.route.params.routeGuide;
         setData(allData);
         setCoord(allData[2]);
-        setRoutes(props.route.params.routeGuide);
         setIsLoading(false);
     }, []);
 
@@ -100,20 +86,19 @@ const Finalized = (props) => {
         );
     } else {
         return (
-            <View style={styles.container}>
-                <View style={styles.header}>
-                    {weatherIcon(weather)}
-                    <TransitRoute routes={routes} />
-                </View>
+            <ScrollView style={styles.container}>
+                <View style={styles.header}>{weatherIcon(weather)}</View>
+                <Text>HI</Text>
+                {/* <TransitRoutes routes={routes} /> */}
                 <View style={styles.body}>
                     <Schedule
                         data={allData}
                         navigation={props.navigation}
                         mapUpdate={mapUpdate}
-                        routeUpdate={routeUpdate}
                         genres={userGenres}
                         accessRights={accessRights}
                         userID={props.userID}
+                        initRoutes={props.route.params.routeGuide}
                     />
 
                     <Modal animated visible={visible} animationType="fade">
@@ -127,7 +112,7 @@ const Finalized = (props) => {
                         <Text style={{ fontSize: 20 }}>Open Map View</Text>
                     </TouchableOpacity>
                 </View>
-            </View>
+            </ScrollView>
         );
     }
 };
@@ -135,13 +120,13 @@ const Finalized = (props) => {
 const styles = StyleSheet.create({
     container: {
         flex: 1,
-        justifyContent: "center",
+        //justifyContent: "center",
     },
     header: {
         flex: 1,
     },
     body: {
-        flex: 4,
+        // flex: 4,
     },
 });
